pyproject.toml
[tool.poetry]
name = "dbt-export"
version = "0.1.0"
description = "Leverage dbt adapters and models to export data to parquet or csv"
authors = ["Alex Butler"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
dbt-core = "^1.0.4"
numpy = "^1.22.3"
pyarrow = "^7.0.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
dbt-sqlite = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
